+++ .metadata
@@ -15,7 +15,7 @@ migration:
     - platform: root
       create_revision: 41456452f29d64e8deb623a3c927524bcf9f111b
       base_revision: 41456452f29d64e8deb623a3c927524bcf9f111b
-    - platform: web
+    - platform: windows
       create_revision: 41456452f29d64e8deb623a3c927524bcf9f111b
       base_revision: 41456452f29d64e8deb623a3c927524bcf9f111b
