+++ android/app/src/main/kotlin/com/example/organiser/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.example.organiser
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}
