+++ lib/views/auth/register.dart
@@ -1,4 +1,4 @@
-import 'package:Organiser/pages/user/add_account_details.dart';
+import 'package:Organiser/views/auth/add_account_details.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
