+++ lib/views/home/home_page.dart
@@ -1,7 +1,7 @@
-import 'package:Organiser/pages/home/screens/socials.dart';
-import 'package:Organiser/pages/home/screens/food.dart';
-import 'package:Organiser/pages/home/screens/schedules.dart';
-import 'package:Organiser/pages/home/screens/tasks.dart';
+import 'package:Organiser/views/home/screens/food.dart';
+import 'package:Organiser/views/home/screens/schedules.dart';
+import 'package:Organiser/views/home/screens/socials.dart';
+import 'package:Organiser/views/home/screens/tasks.dart';
 import 'package:Organiser/widgets/common/bottom_app_bar.dart';
 import 'package:Organiser/widgets/common/drawer.dart';
 import 'package:flutter/material.dart';
