+++ lib/views/info/tips.dart
