+++ lib/widgets/common/app_bar.dart
@@ -1,4 +1,4 @@
-import 'package:Organiser/pages/messages/messages.dart';
+import 'package:Organiser/views/messages/messages.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
