+++ lib/widgets/dialogs/logout.dart
@@ -1,4 +1,4 @@
-import 'package:Organiser/pages/user/login.dart';
+import 'package:Organiser/views/auth/login.dart';
 import 'package:flutter/material.dart';
 import 'package:firebase_auth/firebase_auth.dart';
