+++ lib/widgets/styled/input_adder/location.dart
@@ -1,5 +1,5 @@
 import 'package:Organiser/models/theme/light.dart';
-import 'package:Organiser/providers/theme_provider.dart';
+import 'package:Organiser/services/theme_provider.dart';
 import 'package:flutter/material.dart';
 import 'dart:ui' as ui;
